+++ lib/romp_crm_web/controllers/user_registration_controller.ex
@@ -22,9 +22,7 @@ defmodule RompCrmWeb.UserRegistrationController do
 
     {invitation_registration?, _inv} = invitation_context(conn)
 
-    render(
-      conn,
-      :new,
+    render(conn, :new,
       registration_assigns(
         changeset: changeset,
         invitation_registration: invitation_registration?
@@ -45,14 +43,11 @@ defmodule RompCrmWeb.UserRegistrationController do
 
       {:invitation_email_mismatch, expected, conn} ->
         cs = Accounts.change_user_email(%User{}, attrs)
-
-        flash =
-          "This invitation is for #{expected}. Use that email address or ask for a new invite."
+        flash = "This invitation is for #{expected}. Use that email address or ask for a new invite."
 
         conn
         |> put_flash(:error, flash)
-        |> render(
-          :new,
+        |> render(:new,
           registration_assigns(
             changeset: %{cs | action: :insert},
             invitation_registration: true
@@ -140,9 +135,7 @@ defmodule RompCrmWeb.UserRegistrationController do
         |> redirect(to: ~p"/users/log-in")
 
       {:error, %Ecto.Changeset{} = changeset} ->
-        render(
-          conn,
-          :new,
+        render(conn, :new,
           registration_assigns(
             changeset: %{changeset | action: :insert},
             invitation_registration: true
@@ -156,9 +149,7 @@ defmodule RompCrmWeb.UserRegistrationController do
     changeset = %{cs | action: :insert}
     {_inv?, invitation_registration?} = invitation_context(conn)
 
-    render(
-      conn,
-      :new,
+    render(conn, :new,
       registration_assigns(
         changeset: changeset,
         invitation_registration: invitation_registration?
@@ -235,9 +226,7 @@ defmodule RompCrmWeb.UserRegistrationController do
       {:error, %Ecto.Changeset{} = changeset} ->
         {invitation_registration?, _} = invitation_context(conn)
 
-        render(
-          conn,
-          :new,
+        render(conn, :new,
           registration_assigns(
             changeset: %{changeset | action: :insert},
             invitation_registration: invitation_registration?